==> ntbks/1.save_logs_per_ds.ipynb <==
#mount drive
from google.colab import drive
drive.mount('/content/drive')

cd '/content/drive/MyDrive/TFG-AnalisisResultados'

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import os
import glob

test = pd.read_parquet('./Afina_results_very_new/logs/SingleGP/LD_qqnehvi_run_log0.parquet')

test.head()

# LD data results

# === 1. Folder containing LD logs (SingleGP) ===
folder_path = "./Afina_results_very_new/logs/SingleGP"

# Load all Parquet files starting with 'LD_'
file_paths = glob.glob(os.path.join(folder_path, "LD_*.parquet"))
file_paths = [f for f in file_paths if not os.path.basename(f).startswith("LD_NEURO")]

print(file_paths)

# === 2. Load and concatenate all files ===
all_logs = [pd.read_parquet(fp) for fp in file_paths]
combined_df = pd.concat(all_logs, ignore_index=True)

# Save combined data
combined_df.to_parquet("./LD_SGP_combined_logs.parquet")
# combined_df.to_csv("./LD_combined_logs.csv", index=False)



# === 1. Folder containing LD logs (SingleGP) ===
folder_path = "./Afina_results_very_new/logs/MultiGP"

# Load all Parquet files starting with 'LD_'
file_paths = glob.glob(os.path.join(folder_path, "LD_*.parquet"))
file_paths = [f for f in file_paths if not os.path.basename(f).startswith("LD_NEURO")]

# === 2. Load and concatenate all files ===
all_logs = [pd.read_parquet(fp) for fp in file_paths]
combined_df = pd.concat(all_logs, ignore_index=True)

# Save combined data
combined_df.to_parquet("./LD_MGP_combined_logs.parquet")
# combined_df.to_csv("./LD_combined_logs.csv", index=False)



# === 1. Folder containing LD logs (SingleGP) ===
folder_path = "./Afina_results_very_new/logs/MultiGPR1"

# Load all Parquet files starting with 'LD_'
file_paths = glob.glob(os.path.join(folder_path, "LD_*.parquet"))
file_paths = [f for f in file_paths if not os.path.basename(f).startswith("LD_NEURO")]

# === 2. Load and concatenate all files ===
all_logs = [pd.read_parquet(fp) for fp in file_paths]
combined_df = pd.concat(all_logs, ignore_index=True)

# Save combined data
combined_df.to_parquet("./LD_MGPR1_combined_logs.parquet")
# combined_df.to_csv("./LD_combined_logs.csv", index=False)



df = pd.read_parquet("./Afina_results_very_new/logs/MultiGP/LD_qparego_run_log0.parquet")
df

# LD NEURO data results

# === 1. Folder containing LD logs (SingleGP) ===
folder_path = "./Afina_results_very_new/logs/SingleGP"

# Load all Parquet files starting with 'LD_'
file_paths = glob.glob(os.path.join(folder_path, "LD_NEURO_*.parquet"))

# === 2. Load and concatenate all files ===
all_logs = [pd.read_parquet(fp) for fp in file_paths]
combined_df = pd.concat(all_logs, ignore_index=True)

# Save combined data
combined_df.to_parquet("./LD_NEURO_SGP_combined_logs.parquet")
# combined_df.to_csv("./LD_combined_logs.csv", index=False)



# === 1. Folder containing LD logs (SingleGP) ===
folder_path = "./Afina_results_very_new/logs/MultiGP"

# Load all Parquet files starting with 'LD_'
file_paths = glob.glob(os.path.join(folder_path, "LD_NEURO_*.parquet"))

# === 2. Load and concatenate all files ===
all_logs = [pd.read_parquet(fp) for fp in file_paths]
combined_df = pd.concat(all_logs, ignore_index=True)

# Save combined data
combined_df.to_parquet("./LD_NEURO_MGP_combined_logs.parquet")
# combined_df.to_csv("./LD_combined_logs.csv", index=False)



# === 1. Folder containing LD logs (SingleGP) ===
folder_path = "./Afina_results_very_new/logs/MultiGPR1"

# Load all Parquet files starting with 'LD_'
file_paths = glob.glob(os.path.join(folder_path, "LD_NEURO_*.parquet"))

# === 2. Load and concatenate all files ===
all_logs = [pd.read_parquet(fp) for fp in file_paths]
combined_df = pd.concat(all_logs, ignore_index=True)

# Save combined data
combined_df.to_parquet("./LD_NEURO_MGPR1_combined_logs.parquet")
# combined_df.to_csv("./LD_combined_logs.csv", index=False)



# HD data results

# === 1. Folder containing LD logs (SingleGP) ===
folder_path = "./Afina_results_very_new/logs/SingleGP"

# Load all Parquet files starting with 'LD_'
file_paths = glob.glob(os.path.join(folder_path, "HD_*.parquet"))

# === 2. Load and concatenate all files ===
all_logs = [pd.read_parquet(fp) for fp in file_paths]
combined_df = pd.concat(all_logs, ignore_index=True)

# Save combined data
combined_df.to_parquet("./HD_SGP_combined_logs.parquet")
# combined_df.to_csv("./LD_combined_logs.csv", index=False)



# === 1. Folder containing LD logs (SingleGP) ===
folder_path = "./Afina_results_very_new/logs/MultiGP"

# Load all Parquet files starting with 'LD_'
file_paths = glob.glob(os.path.join(folder_path, "HD_*.parquet"))

# === 2. Load and concatenate all files ===
all_logs = [pd.read_parquet(fp) for fp in file_paths]
combined_df = pd.concat(all_logs, ignore_index=True)

# Save combined data
combined_df.to_parquet("./HD_MGP_combined_logs.parquet")
# combined_df.to_csv("./LD_combined_logs.csv", index=False)



# === 1. Folder containing LD logs (SingleGP) ===
folder_path = "./Afina_results_very_new/logs/MultiGPR1"

# Load all Parquet files starting with 'LD_'
file_paths = glob.glob(os.path.join(folder_path, "HD_*.parquet"))

# === 2. Load and concatenate all files ===
all_logs = [pd.read_parquet(fp) for fp in file_paths]
combined_df = pd.concat(all_logs, ignore_index=True)

# Save combined data
combined_df.to_parquet("./HD_MGPR1_combined_logs.parquet")
# combined_df.to_csv("./LD_combined_logs.csv", index=False)
