==> ntbks/4.create_additional_stats.ipynb <==
#mount drive
from google.colab import drive
drive.mount('/content/drive')

cd '/content/drive/MyDrive/TFG-AnalisisResultados'

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import os
import glob

#LD

for df_path in ['./LD_SGP_combined_logs.parquet', './LD_MGP_combined_logs.parquet', './LD_MGPR1_combined_logs.parquet']:
  df = pd.read_parquet(df_path)
  summary_records = []

  # === 3. Group by acquisition and seed (i.e., per run) ===
  grouped = df.groupby(['acquisition', 'seed'])


  # For each acquisition + seed/run
  for (acq, seed), run_df in grouped:
      run_df = run_df.sort_values('iteration')
      last_iter = run_df['iteration'].max()

      final_front = run_df.loc[run_df['iteration'] == last_iter]

      # Assuming objectives is a list of shape [num_points, 2] where [:, 0] = accuracy, [:, 1] = depth
      final_obj = np.stack(final_front['pareto_front'].values[0])  # shape (batch_size, 2)

      acc_mean = np.mean(final_obj[:, 0])
      acc_std = np.std(final_obj[:, 0])
      depth_mean = np.mean(final_obj[:, 1])
      depth_std = np.std(final_obj[:, 1])

      mindepth_index = np.argmax(final_obj[:, 1])
      maxacc_index = np.argmax(final_obj[:, 0])
      # Extract the corresponding row
      min_depth_pair = final_obj[mindepth_index]
      max_accuracy_pair = final_obj[maxacc_index]

      # Drop rows without posterior std (i.e., Sobol)
      if acq != "Sobol":
          std_vals = np.stack(run_df['posterior_std'].dropna().values)  # list of arrays (batch_size x 2)

          # Collect per-objective stds per iteration
          obj0_stds = []
          obj1_stds = []

          for arr in std_vals:
              arr = np.stack(arr)  # shape (batch_size, 2)
              obj0_stds.append(arr[:, 0].mean())  # mean over batch, objective 0
              obj1_stds.append(arr[:, 1].mean())  # mean over batch, objective 1

          # Mean across iterations
          posterior_std_obj0_mean = np.mean(obj0_stds)
          posterior_std_obj1_mean = np.mean(obj1_stds)
          posterior_std_obj0_std = np.std(obj0_stds)
          posterior_std_obj1_std = np.std(obj1_stds)
      else:
          posterior_std_obj0_mean = posterior_std_obj1_mean = np.nan
          posterior_std_obj0_std = posterior_std_obj1_std = np.nan


      summary_records.append({
          "acquisition": acq,
          "seed": seed,
          "pareto_acc_mean": acc_mean,
          "pareto_acc_std": acc_std,
          "pareto_depth_mean": depth_mean,
          "pareto_depth_std": depth_std,
          "min_depth_pair": min_depth_pair,
          "max_accuracy_pair": max_accuracy_pair,
          "posterior_std_obj0_mean": posterior_std_obj0_mean,
          "posterior_std_obj0_std": posterior_std_obj0_std,
          "posterior_std_obj1_mean": posterior_std_obj1_mean,
          "posterior_std_obj1_std": posterior_std_obj1_std,

      })

  per_run_summary = pd.DataFrame(summary_records)

  # === 4. Aggregate Stats per Acquisition Function ===
  final_summary = per_run_summary.groupby('acquisition').agg({
      'pareto_acc_mean': ['mean', 'std'],
      'pareto_acc_std': ['mean'],  # Optional: std of stds
      'pareto_depth_mean': ['mean', 'std'],
      'pareto_depth_std': ['mean'],  # Optional
      'min_depth_pair': list,
      'max_accuracy_pair': list,
      'posterior_std_obj0_mean': ['mean', 'std'],
      'posterior_std_obj0_std': ['mean'],
      'posterior_std_obj1_mean': ['mean', 'std'],
      'posterior_std_obj1_std': ['mean'],

  }).reset_index()

  # Flatten column names
  final_summary.columns = ['_'.join(col).strip('_') for col in final_summary.columns]

  final_summary['best_min_depth_pair'] = final_summary['min_depth_pair_list'].apply(
    lambda lst: max(lst, key=lambda x: x[1]) if lst else np.nan
  )

  final_summary['best_max_accuracy_pair'] = final_summary['max_accuracy_pair_list'].apply(
    lambda lst: max(lst, key=lambda x: x[0]) if lst else np.nan
  )

  # === 5. Save summary ===
  parts = df_path.split('_')
  name = '_'.join(parts[:2])

  # Table pareto front
  obj_cols = ['acquisition', 'pareto_acc_mean_mean', 'pareto_acc_mean_std',
              'pareto_depth_mean_mean', 'pareto_depth_mean_std', 'best_min_depth_pair', 'best_max_accuracy_pair']
  final_summary[obj_cols].to_csv(name + "_final_objectives.csv", index=False)

  # Table gp uncertainty
  unc_cols = ['acquisition', 'posterior_std_obj0_mean_mean', 'posterior_std_obj0_mean_std', 'posterior_std_obj1_mean_mean', 'posterior_std_obj1_mean_std']
  final_summary[unc_cols].to_csv(name + "_gp_uncertainty.csv", index=False)


final_summary.head()

#LD_NEURO

for df_path in ['./LD_NEURO_SGP_combined_logs.parquet', './LD_NEURO_MGP_combined_logs.parquet', './LD_NEURO_MGPR1_combined_logs.parquet']:
  df = pd.read_parquet(df_path)
  summary_records = []

  # === 3. Group by acquisition and seed (i.e., per run) ===
  grouped = df.groupby(['acquisition', 'seed'])


  # For each acquisition + seed/run
  for (acq, seed), run_df in grouped:
      run_df = run_df.sort_values('iteration')
      last_iter = run_df['iteration'].max()

      final_front = run_df.loc[run_df['iteration'] == last_iter]

      # Assuming objectives is a list of shape [num_points, 2] where [:, 0] = accuracy, [:, 1] = depth
      final_obj = np.stack(final_front['pareto_front'].values[0])  # shape (batch_size, 2)

      acc_mean = np.mean(final_obj[:, 0])
      acc_std = np.std(final_obj[:, 0])
      depth_mean = np.mean(final_obj[:, 1])
      depth_std = np.std(final_obj[:, 1])

      mindepth_index = np.argmax(final_obj[:, 1])
      maxacc_index = np.argmax(final_obj[:, 0])
      # Extract the corresponding row
      min_depth_pair = final_obj[mindepth_index]
      max_accuracy_pair = final_obj[maxacc_index]

      # Drop rows without posterior std (i.e., Sobol)
      if acq != "Sobol":
          std_vals = np.stack(run_df['posterior_std'].dropna().values)  # list of arrays (batch_size x 2)

          # Collect per-objective stds per iteration
          obj0_stds = []
          obj1_stds = []

          for arr in std_vals:
              arr = np.stack(arr)  # shape (batch_size, 2)
              obj0_stds.append(arr[:, 0].mean())  # mean over batch, objective 0
              obj1_stds.append(arr[:, 1].mean())  # mean over batch, objective 1

          # Mean across iterations
          posterior_std_obj0_mean = np.mean(obj0_stds)
          posterior_std_obj1_mean = np.mean(obj1_stds)
          posterior_std_obj0_std = np.std(obj0_stds)
          posterior_std_obj1_std = np.std(obj1_stds)
      else:
          posterior_std_obj0_mean = posterior_std_obj1_mean = np.nan
          posterior_std_obj0_std = posterior_std_obj1_std = np.nan


      summary_records.append({
          "acquisition": acq,
          "seed": seed,
          "pareto_acc_mean": acc_mean,
          "pareto_acc_std": acc_std,
          "pareto_depth_mean": depth_mean,
          "pareto_depth_std": depth_std,
          "min_depth_pair": min_depth_pair,
          "max_accuracy_pair": max_accuracy_pair,
          "posterior_std_obj0_mean": posterior_std_obj0_mean,
          "posterior_std_obj0_std": posterior_std_obj0_std,
          "posterior_std_obj1_mean": posterior_std_obj1_mean,
          "posterior_std_obj1_std": posterior_std_obj1_std,

      })

  per_run_summary = pd.DataFrame(summary_records)

  # === 4. Aggregate Stats per Acquisition Function ===
  final_summary = per_run_summary.groupby('acquisition').agg({
      'pareto_acc_mean': ['mean', 'std'],
      'pareto_acc_std': ['mean'],  # Optional: std of stds
      'pareto_depth_mean': ['mean', 'std'],
      'pareto_depth_std': ['mean'],  # Optional
      'min_depth_pair': list,
      'max_accuracy_pair': list,
      'posterior_std_obj0_mean': ['mean', 'std'],
      'posterior_std_obj0_std': ['mean'],
      'posterior_std_obj1_mean': ['mean', 'std'],
      'posterior_std_obj1_std': ['mean'],

  }).reset_index()

  # Flatten column names
  final_summary.columns = ['_'.join(col).strip('_') for col in final_summary.columns]

  final_summary['best_min_depth_pair'] = final_summary['min_depth_pair_list'].apply(
    lambda lst: max(lst, key=lambda x: x[1]) if lst else np.nan
  )

  final_summary['best_max_accuracy_pair'] = final_summary['max_accuracy_pair_list'].apply(
    lambda lst: max(lst, key=lambda x: x[0]) if lst else np.nan
  )

  # === 5. Save summary ===
  parts = df_path.split('_')
  name = '_'.join(parts[:3])

  # Table pareto front
  obj_cols = ['acquisition', 'pareto_acc_mean_mean', 'pareto_acc_mean_std',
              'pareto_depth_mean_mean', 'pareto_depth_mean_std', 'best_min_depth_pair', 'best_max_accuracy_pair']
  final_summary[obj_cols].to_csv(name + "_final_objectives.csv", index=False)

  # Table gp uncertainty
  unc_cols = ['acquisition', 'posterior_std_obj0_mean_mean', 'posterior_std_obj0_mean_std', 'posterior_std_obj1_mean_mean', 'posterior_std_obj1_mean_std']
  final_summary[unc_cols].to_csv(name + "_gp_uncertainty.csv", index=False)


final_summary.head()



#HD

for df_path in ['./HD_SGP_combined_logs.parquet', './HD_MGP_combined_logs.parquet', './HD_MGPR1_combined_logs.parquet']:
  df = pd.read_parquet(df_path)
  summary_records = []

  # === 3. Group by acquisition and seed (i.e., per run) ===
  grouped = df.groupby(['acquisition', 'seed'])


  # For each acquisition + seed/run
  for (acq, seed), run_df in grouped:
      run_df = run_df.sort_values('iteration')
      last_iter = run_df['iteration'].max()

      final_front = run_df.loc[run_df['iteration'] == last_iter]

      # Assuming objectives is a list of shape [num_points, 2] where [:, 0] = accuracy, [:, 1] = depth
      final_obj = np.stack(final_front['pareto_front'].values[0])  # shape (batch_size, 2)

      acc_mean = np.mean(final_obj[:, 0])
      acc_std = np.std(final_obj[:, 0])
      depth_mean = np.mean(final_obj[:, 1])
      depth_std = np.std(final_obj[:, 1])

      mindepth_index = np.argmax(final_obj[:, 1])
      maxacc_index = np.argmax(final_obj[:, 0])
      # Extract the corresponding row
      min_depth_pair = final_obj[mindepth_index]
      max_accuracy_pair = final_obj[maxacc_index]

      # Drop rows without posterior std (i.e., Sobol)
      if acq != "Sobol":
          std_vals = np.stack(run_df['posterior_std'].dropna().values)  # list of arrays (batch_size x 2)

          # Collect per-objective stds per iteration
          obj0_stds = []
          obj1_stds = []

          for arr in std_vals:
              arr = np.stack(arr)  # shape (batch_size, 2)
              obj0_stds.append(arr[:, 0].mean())  # mean over batch, objective 0
              obj1_stds.append(arr[:, 1].mean())  # mean over batch, objective 1

          # Mean across iterations
          posterior_std_obj0_mean = np.mean(obj0_stds)
          posterior_std_obj1_mean = np.mean(obj1_stds)
          posterior_std_obj0_std = np.std(obj0_stds)
          posterior_std_obj1_std = np.std(obj1_stds)
      else:
          posterior_std_obj0_mean = posterior_std_obj1_mean = np.nan
          posterior_std_obj0_std = posterior_std_obj1_std = np.nan


      summary_records.append({
          "acquisition": acq,
          "seed": seed,
          "pareto_acc_mean": acc_mean,
          "pareto_acc_std": acc_std,
          "pareto_depth_mean": depth_mean,
          "pareto_depth_std": depth_std,
          "min_depth_pair": min_depth_pair,
          "max_accuracy_pair": max_accuracy_pair,
          "posterior_std_obj0_mean": posterior_std_obj0_mean,
          "posterior_std_obj0_std": posterior_std_obj0_std,
          "posterior_std_obj1_mean": posterior_std_obj1_mean,
          "posterior_std_obj1_std": posterior_std_obj1_std,

      })

  per_run_summary = pd.DataFrame(summary_records)

  # === 4. Aggregate Stats per Acquisition Function ===
  final_summary = per_run_summary.groupby('acquisition').agg({
      'pareto_acc_mean': ['mean', 'std'],
      'pareto_acc_std': ['mean'],  # Optional: std of stds
      'pareto_depth_mean': ['mean', 'std'],
      'pareto_depth_std': ['mean'],  # Optional
      'min_depth_pair': list,
      'max_accuracy_pair': list,
      'posterior_std_obj0_mean': ['mean', 'std'],
      'posterior_std_obj0_std': ['mean'],
      'posterior_std_obj1_mean': ['mean', 'std'],
      'posterior_std_obj1_std': ['mean'],

  }).reset_index()

  # Flatten column names
  final_summary.columns = ['_'.join(col).strip('_') for col in final_summary.columns]

  final_summary['best_min_depth_pair'] = final_summary['min_depth_pair_list'].apply(
    lambda lst: max(lst, key=lambda x: x[1]) if lst else np.nan
  )

  final_summary['best_max_accuracy_pair'] = final_summary['max_accuracy_pair_list'].apply(
    lambda lst: max(lst, key=lambda x: x[0]) if lst else np.nan
  )

  # === 5. Save summary ===
  parts = df_path.split('_')
  name = '_'.join(parts[:2])

  # Table pareto front
  obj_cols = ['acquisition', 'pareto_acc_mean_mean', 'pareto_acc_mean_std',
              'pareto_depth_mean_mean', 'pareto_depth_mean_std', 'best_min_depth_pair', 'best_max_accuracy_pair']
  final_summary[obj_cols].to_csv(name + "_final_objectives.csv", index=False)

  # Table gp uncertainty
  unc_cols = ['acquisition', 'posterior_std_obj0_mean_mean', 'posterior_std_obj0_mean_std', 'posterior_std_obj1_mean_mean', 'posterior_std_obj1_mean_std']
  final_summary[unc_cols].to_csv(name + "_gp_uncertainty.csv", index=False)


final_summary.head()
