==> ntbks/5.Analyze_adittional_stats.ipynb <==
#mount drive
from google.colab import drive
drive.mount('/content/drive')

cd '/content/drive/MyDrive/TFG-AnalisisResultados'

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import os
import glob

# LD

pd.set_option('display.max_columns', None)
sgp = pd.read_csv('./LD_SGP_final_objectives.csv')
sgp

for df_path in ['./LD_SGP_final_objectives.csv', './LD_MGP_final_objectives.csv', './LD_MGPR1_final_objectives.csv']:

  print(f"\n###########  Table for {df_path}  ############\n")

  table8 = pd.read_csv(df_path)
  # Format as "mean ± std"
  table8['Pareto Accuracy'] = table8.apply(
      lambda row: f"{row['pareto_acc_mean_mean']:.3f} ± {row['pareto_acc_mean_std']:.3f}", axis=1)
  table8['Pareto Depth'] = table8.apply(
      lambda row: f"{row['pareto_depth_mean_mean']:.3f} ± {row['pareto_depth_mean_std']:.3f}", axis=1)

  # Select and rename
  table8 = table8[['acquisition', 'Pareto Accuracy', 'Pareto Depth']]

  table8.columns = ['Acquisition', 'Pareto Accuracy', 'Pareto Depth']
  print(table8.to_latex(index=False))


for df_path in ['./LD_SGP_final_objectives.csv', './LD_MGP_final_objectives.csv', './LD_MGPR1_final_objectives.csv']:
  table = pd.read_csv(df_path)

  print(f"\n#######  Table 1 for {df_path}  ########\n")
  print(table['best_min_depth_pair'].to_latex())
  print(f"\n#######  Table 2 for {df_path}  #######\n")
  print(table['best_max_accuracy_pair'].to_latex())

# LD NEURO

pd.set_option('display.max_columns', None)
sgp = pd.read_csv('./LD_NEURO_SGP_final_objectives.csv')
sgp

for df_path in ['./LD_NEURO_SGP_final_objectives.csv', './LD_NEURO_MGP_final_objectives.csv', './LD_NEURO_MGPR1_final_objectives.csv']:

  print(f"\n###########  Table for {df_path}  ############\n")

  table8 = pd.read_csv(df_path)
  # Format as "mean ± std"
  table8['Pareto Accuracy'] = table8.apply(
      lambda row: f"{row['pareto_acc_mean_mean']:.3f} ± {row['pareto_acc_mean_std']:.3f}", axis=1)
  table8['Pareto Depth'] = table8.apply(
      lambda row: f"{row['pareto_depth_mean_mean']:.3f} ± {row['pareto_depth_mean_std']:.3f}", axis=1)

  # Select and rename
  table8 = table8[['acquisition', 'Pareto Accuracy', 'Pareto Depth']]

  table8.columns = ['Acquisition', 'Pareto Accuracy', 'Pareto Depth']
  print(table8.to_latex(index=False))


for df_path in ['./LD_NEURO_SGP_final_objectives.csv', './LD_NEURO_MGP_final_objectives.csv', './LD_NEURO_MGPR1_final_objectives.csv']:
  table = pd.read_csv(df_path)

  print(f"\n#######  Table 1 for {df_path}  ########\n")
  print(table['best_min_depth_pair'].to_latex())
  print(f"\n#######  Table 2 for {df_path}  #######\n")
  print(table['best_max_accuracy_pair'].to_latex())

# HD

pd.set_option('display.max_columns', None)
sgp = pd.read_csv('./HD_SGP_final_objectives.csv')
sgp

for df_path in ['./HD_SGP_final_objectives.csv', './HD_MGP_final_objectives.csv', './HD_MGPR1_final_objectives.csv']:

  print(f"\n###########  Table for {df_path}  ############\n")

  table8 = pd.read_csv(df_path)
  # Format as "mean ± std"
  table8['Pareto Accuracy'] = table8.apply(
      lambda row: f"{row['pareto_acc_mean_mean']:.3f} ± {row['pareto_acc_mean_std']:.3f}", axis=1)
  table8['Pareto Depth'] = table8.apply(
      lambda row: f"{row['pareto_depth_mean_mean']:.3f} ± {row['pareto_depth_mean_std']:.3f}", axis=1)

  # Select and rename
  table8 = table8[['acquisition', 'Pareto Accuracy', 'Pareto Depth']]

  table8.columns = ['Acquisition', 'Pareto Accuracy', 'Pareto Depth']
  print(table8.to_latex(index=False))


for df_path in ['./HD_SGP_final_objectives.csv', './HD_MGP_final_objectives.csv', './HD_MGPR1_final_objectives.csv']:
  table = pd.read_csv(df_path)

  print(f"\n#######  Table 1 for {df_path}  ########\n")
  print(table['best_min_depth_pair'].to_latex())
  print(f"\n#######  Table 2 for {df_path}  #######\n")
  print(table['best_max_accuracy_pair'].to_latex())
